+++ src/main/kotlin/com/tezos/lang/michelson/structureView/MichelsonCreateContractViewElement.kt
@@ -1,0 +1,18 @@
+package com.tezos.lang.michelson.structureView
+
+import com.intellij.ide.structureView.StructureViewTreeElement
+import com.intellij.ide.structureView.impl.common.PsiTreeElementBase
+import com.tezos.lang.michelson.psi.PsiCreateContractInstruction
+import java.util.*
+
+internal class MichelsonCreateContractViewElement(private val instruction: PsiCreateContractInstruction) : PsiTreeElementBase<PsiCreateContractInstruction>(instruction) {
+    override fun getPresentableText(): String = "Create contract"
+
+    override fun getChildrenBase(): MutableCollection<StructureViewTreeElement> {
+        val contract = instruction.contract
+        return when (contract) {
+            null -> Collections.emptyList()
+            else -> mutableListOf(MichelsonContractViewElement(contract))
+        }
+    }
+}